% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_shape_circle2}
\alias{add_shape_circle2}
\title{Add additional igraph vertex shape}
\usage{
add_shape_circle2()
}
\description{
a circle with variable frame border width (not available in the standard shapes)
}
