% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_ngrams.R
\name{bind_ngrams}
\alias{bind_ngrams}
\title{Replace blanks by replacement pattern in known ngrams in a string}
\usage{
bind_ngrams(string, ngrams, replacement = "_", case_insensitive = TRUE)
}
\arguments{
\item{string}{A character vector in which blanks of ngrams shall be replaced.}

\item{ngrams}{Character vector of known ngrams. Please note that ngrams in the \code{return} will have the case formatting of these \code{ngrams}.}

\item{case_insensitive}{By default TRUE. Note that case is only used for matching (see ngram parameter)}
}
\value{
The \code{string} with modified ngrams.
}
\description{
Usually ngrams are identified and modified by probabilistic collocation extraction,
but in certain situations one might want to fix specific word combinations before further processing of the text
independent of collocation statistics such as PMI.
}
\examples{

bind_ngrams(c("The United Nations are an important organization.",
              "They are concerned, e.g., with sustainable development and climate change.")
            , ngrams = c("United Nations", "CLIMATE CHANGE"))
# [1] "The United_Nations are an important organization."
# [2] "They are concerned, e.g., with sustainable development and CLIMATE_CHANGE."
}
