% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cooccurrence.R
\name{get_cooccurrence}
\alias{get_cooccurrence}
\title{Get the co-occurrence of column elements per row}
\usage{
get_cooccurrence(m, cols, format_output = as.matrix)
}
\arguments{
\item{format_output}{}
}
\description{
Documentation to be done....
}
\examples{
mat <- matrix(sample(1:2, 18, replace = T), nrow = 3)
get_cooccurrence(m = mat, c(1,3,6))
#     [,1] [,2] [,3]
# [1,]    3    3    3
# [2,]    3    3    3
# [3,]    3    3    3
get_cooccurrence(m = mat, c(1,3,6), format_output = as.simple_triplet_matrix)
#A 3x3 simple triplet matrix.
}
