% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loglik_parallel.R
\name{get_loglik_parallel}
\alias{get_loglik_parallel}
\title{Get loglikelihood from multiple locally saved models}
\usage{
get_loglik_parallel(modelfiles, ncores)
}
\arguments{
\item{modelfiles}{The full filenames as \code{character} of the models saved as \code{.rds} to extract loglik.}

\item{ncores}{The number of cores to be used.
Advise: leave one processor spare, if you want to
use your computer during operation of the function.}
}
\value{
A \code{data.table} listing the number of topics of each model with associated loglik.
}
\description{
The function loads each of the specified LDA models created with text2vec
in parallel and extracts the loglikelihodd from the last iteration step.
This function is useful if not all models can be loaded into RAM at once 
due to memory restrictions.
}
