% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_skip_gram_windows.R
\name{get_n_skip_gram_windows}
\alias{get_n_skip_gram_windows}
\title{Get the number of skip gram windows}
\usage{
get_n_skip_gram_windows(tokens, window_size)
}
\arguments{
\item{tokens}{A list of tokens, as created, e.g., by \code{text2vec::wordtokenizer()}.}

\item{window_size}{The window size to be considered.}
}
\value{
The total number of windows or virtual documents.
}
\description{
Given a list of tokens, the functions calculates the number of skip gram windows
that is used, e.g., during creating a \code{tcm} with \code{text2vec}, for a given window size. This function
calcualtes something like the total number of virtual documents used to create a \code{tcm}.
}
