% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topic_trends.R
\name{get_topic_trends}
\alias{get_topic_trends}
\title{Model and summarize topic trends based on text2vec LDA}
\usage{
get_topic_trends(doc_topic_distr, top_terms, doc_ids_years,
  loess_aicc_span = c(0.1, 0.9), sep_labels = "#",
  topic_selection_qt = 0.95, topic_selection_top_n = 10,
  topic_selection_pattern)
}
\arguments{
\item{topic_selection_pattern}{}
}
\description{
Model and summarize topic trends based on text2vec LDA
}
\examples{

}
