% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gradient_numeric}
\alias{gradient_numeric}
\title{Get numeric gradient between points}
\usage{
gradient_numeric(x, y)
}
\arguments{
\item{x}{steps / independent variable}

\item{y}{observed data points / response variable}
}
\value{
Matrix showing the gradient values between points.
        The x positions of gradient values are returned as the middle point between the steps.
        Note that \code{nrow(gradient) = length(x)-1}.
}
\description{
Simple caluclation of gradient between points.
}
