% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess_aicc_optimized.R
\name{loess_aicc_optimized}
\alias{loess_aicc_optimized}
\title{Fitting of AICC optimized loess per topic over specified time range}
\usage{
loess_aicc_optimized(DT, topic_selection = NULL, year_selection = NULL,
  fit_columns = NULL, span_interval = NULL)
}
\arguments{
\item{DT}{A \code{data.table}. Current implementation requires data.table to have two columns named year and topic.}

\item{topic_selection}{By default \code{loess} is fitted for all topics. A selection of topic ids may be specified here.}

\item{year_selection}{By default \code{loess} is fitted for all years. A selection of years may be specified here.}

\item{fit_columns}{The columns to be fitted by \code{loess}.}

\item{span_interval}{The span intervall to be considered for optimization of loess span.}
}
\value{
The original \code{DT} with additional columns as specified via \code{fit_columns} with the suffix \code{"_loess"}.
         The entires in the new columns for combinations of year and topic that have not specified in
         the \code{topic_selection} or \code{year_selection} are set to \code{NA} (standard DT join).
}
\description{
Given a data.table with a year and topic column and additional columns with values for these variables (e.g. mean topic probability per year),
the function fits a loess model and optimizes the span via Akaikes information criterion for each topic.
Modelling may be limited to a specified time range.
The results are joined with the input data.table. NOTE: currently the order of columns is not preserved.
}
