% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_lm.R
\name{p_lm}
\alias{p_lm}
\title{Get p values of linear model}
\usage{
p_lm(model)
}
\arguments{
\item{model}{A linear model fitted via \code{lm}.}
}
\value{
Named numeric vector containing p-values of the model, i.e., the overall p-value of the F-statistic,
        and the p-values of the intercept and coefficients.
        For naming convention of output see example.
}
\description{
Get p values of linear model
}
\examples{

x = rnorm(42)
y = x + rnorm(42)
m = lm(y ~ x)
p_lm(m)
#      p_model  p_intercept          p_x
# 9.010565e-11 3.330814e-01 9.010565e-11
}
