% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns_in_topics.R
\name{patterns_in_topics}
\alias{patterns_in_topics}
\title{Check occurrence and rank of patterns per topic}
\usage{
patterns_in_topics(term_patterns, topics, check_n_top_terms = 50)
}
\arguments{
\item{term_patterns}{The term patterns to be checked for each topic.
Passed to \code{grep}, hence, regex such as "|" for checking multiple terms at a time allowed.}

\item{topics}{A character \code{matrix} containing term lists per topic.
The input might e.g. be \code{topicmodels::terms(model, model@Dim[2])}}

\item{check_n_top_terms}{The number of top terms to be considered for the check.}
}
\value{
A list of character matrices named after and having elements equal to number of \code{term_patterns}.
        Matrices contain the input \code{topics} reduced to the columns that contain the pattern to be checked
        and reduced to number of rows as specified via \code{check_n_top_terms}.
        Furthermore, the minimum rank (1 is best) of the pattern in each topic is pasted into the respective colname.
}
\description{
Check occurrence and rank of patterns per topic
}
\examples{



patterns_in_topics(term_patterns = c("environ*", "terms")
                  , topics = cbind(c("irrelevant", "topic", "terms"), c("global", "environmental", "protection"))
                  , check_n_top_terms = 2)
$`environ*`
Topic 2 - min_rank_2
[1,] "global"
[2,] "environmental"

$terms
NULL
}
