% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_topic_trends.R
\name{plot_topic_trends}
\alias{plot_topic_trends}
\title{Generate trend plots for topics modelled with text2vec and processed with textility}
\usage{
plot_topic_trends(topic_trends, topic_selection, n = NULL,
  exclude_topics = NULL, out_dir = NULL, filename_suffix = "",
  ylim = NULL, facet_columns = 3, text_size = 12, device = "pdf")
}
\arguments{
\item{topic_trends}{...}

\item{topic_selection}{Named list of character vectors with the topic numbers to be plotted (and stored) under the list name.}

\item{n}{Number of top topics to plot per list in \code{topic_selection}.}

\item{exclude_topics}{A character vector giving the topic numbers to be excluded from all lists in \code{topic_selection}.}

\item{out_dir}{Directory where plots are saved.}

\item{filename_suffix}{A string to specify the end of the filename inserted before \code{.filetype}.}

\item{ylim}{The y limits, a numeric vector with two entries. By default \code{c(0, max(prob_mean))}.}

\item{facet_columns}{The number of columns to be used for the facet plot.}

\item{text_size}{By default \code{12}.}

\item{device}{The plotting device (output filetype), by default \code{"pdf"}.}
}
\description{
Documentation to be updated...
}
