% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_acronyms_by_words.R
\name{replace_acronyms_by_words}
\alias{replace_acronyms_by_words}
\title{Replace acronyms by words}
\usage{
replace_acronyms_by_words(docs, remove_replaced_acr = TRUE,
  remove_all_bracketed_acr = TRUE)
}
\arguments{
\item{docs}{The documents to be processed in form of a \code{character} vector. Each document is processed separately, hence no shared acronyms are assumed.}

\item{remove_replaced_acr}{By default TRUE. If an acronym was identified and replaced, the acronym itself is deleted.}

\item{remove_all_bracketed_acr}{By default TRUE. All acronyms in brackets are removed. The identification pattern is \code{([A-Z][A-Za-z]*[A-Z]s?\\)}.}
}
\value{
The documents wit acronyms replaced.
}
\description{
Searches for acronyms that are followed by a bracketed explanation and replaces the acronym by the content in the brackets.
}
\details{
Currently requires stringi and magrittr loaded (the latter dependency will be removed in the future).

NOTE: The implementation is currently quite slow if there are many acronyms that are not matched. Might be improved in future...

Documentation still to be updated, please check the code in the meantime.
}
\examples{
replace_acronyms_by_words("The United Nations Framework Convention on Climate Change (UNFCCC) is a milestone. The UNFCCC came into force in 1994.")

}
