% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_columns.R
\name{split_columns}
\alias{split_columns}
\title{Split columns on basis of a string and duplicate entries}
\usage{
split_columns(x, pattern)
}
\arguments{
\item{x}{A \code{data.frame} or \code{data.table}.}

\item{pattern}{The pattern to split the columns (will not be included in the new column names).}
}
\value{
Dependding on the input class, a \code{data.frame} or \code{data.table} with the additional columns.
}
\description{
Based on \code{stringi::stri_split_fixed()} column names are split on basis of a split pattern into new columns.
The content of the original column is duplicated in all split columns.
}
\examples{
df <- data.frame(split_here = 11:13, and_here = letters[11:13],  and_again_here = 14:16)
split_columns(df, "_")
#   split here and here and again here
# 1    11   11   k    k  14    14   14
# 2    12   12   l    l  15    15   15
# 3    13   13   m    m  16    16   16
}
