% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_rows_DT.R
\name{subset_rows_DT}
\alias{subset_rows_DT}
\title{Subset rows of a data.table by reference}
\usage{
subset_rows_DT(DT, idxs, delete_idxs = FALSE)
}
\arguments{
\item{DT}{The \code{data.table} to be subsetted.}

\item{idxs}{An \code{integer} vector with row indices for the subset. For a conditional subsets see example below.}

\item{del.idxs}{FALSE by default to get a positive and return the selected rows. Set to FALSE for deleting rows.}
}
\value{
A positive (or negative) subset of the input \code{data.table}.
}
\description{
Subset rows of a data.table by reference
}
\examples{
#conditional deletion of rows
DT = data.table(x=rep(c("b","a","c"),each=3), y=c(1,3,6), v=1:9)
#introduce numbering id if relevant
DT[, id := 1:nrow(DT)]
delete_letters <- c("a", "c")
delete_idxs <- which(DT[,x] \%in\% delete_letters)
DT_subset <- subset_rows_DT(DT = DT, idxs = delete_idxs, delete_idxs = TRUE)
DT_subset
#    x y v id
# 1: b 1 1  1
# 2: b 3 2  2
# 3: b 6 3  3
}
