% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tcm_specs_standard}
\alias{tcm_specs_standard}
\title{Initializes the standard specifications for tcms to be used as reference for coherence calculation}
\usage{
tcm_specs_standard()
}
\value{
data.table with standard specifications, see example...
}
\description{
Initializes the standard specifications for tcms to be used as reference for coherence calculation
}
\examples{
tcm_specs_standard()
#            tcm_id type window_size
# 1: standard_ref_1  ext           5
# 2: standard_ref_2  ext          10
# 3: standard_ref_3  ext         110
# 4: standard_ref_4  int         Inf
}
