% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_cooccurrence.R
\name{topic_cooccurrence}
\alias{topic_cooccurrence}
\title{Calculate topic co-occurrence from text2vec model}
\usage{
topic_cooccurrence(x, n_top_words = 5, term_order_lambda = 1,
  diag_to_zero = TRUE, n_top_topics = 10)
}
\arguments{
\item{x}{A named list containing a fitted text2vec LDA model and the document topic distribution.
Names need to be: \code{list(model = ..., doc_topic_distr = ...)}.
The first element should be the lda model object and the second the topic distribution
to be extraced from the former via \code{$doc_topic_distr}.}

\item{n_top_words}{Number of top words to of each topic to be used as topic label. Passed to \code{get_top_words}.}

\item{term_order_lambda}{Lambda for re-ordering the extracted top terms. Passed to \code{get_top_words}.}

\item{diag_to_zero}{Shall the diagonal of the output be set to zero, i.e.,
should the self-co-occurrence (which is the occurrence of a topic of all docs) be neglected?
By default \code{TRUE}.#'}
}
\value{
An symmetric matrix showing the co-occurrence of topics in documents.
}
\description{
Calculate topic co-occurrence from text2vec model
}
