% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncontract_negations.R
\name{uncontract_negations}
\alias{uncontract_negations}
\title{Uncontrtact negations, e.g., "don't" to "do not"}
\usage{
uncontract_negations(x, negation_forms = "default")
}
\arguments{
\item{x}{The \code{string} whose negations shall be uncontracted.}

\item{negation_forms}{A \code{matrix} with two columns of mode \code{character}
The first column contains the contracted form, the second the uncontracted form.
The default is "default", which considers the following negations:
"isn.t","aren.t","wasn.t","weren.t","hasn.t"  , "haven.t",
"hadn.t"  , "doesn.t" , "don.t" ,  "didn.t" , "won.t" ,  "wouldn.t",
"shan.t" ,  "shouldn.t" ,"can.t" ,  "couldn.t" ,"mustn.t" , "cannot"}
}
\value{
The \code{string} with uncontracted negations.
}
\description{
Uncontrtact negations, e.g., "don't" to "do not"
}
\examples{
txt <- c("I don`t want this.",  "I won`t accept it.")
uncontract_negations(txt)
#[1] "I do not want this."   "I will not accept it."
}
