% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_coherence.R
\name{calc_coherence}
\alias{calc_coherence}
\title{Calculation of various coherence measures for topic models based on LDA algorithm}
\usage{
calc_coherence(dtm, beta, n = 10)
}
\arguments{
\item{dtm}{The term document matrix, may be a \code{sparseMatrix} or \code{simple_triplet_matrix}.}

\item{beta}{Either the raw distribution of word probabilities (columnnames = words, entries = probabiliities) over topics (each row = Topic), e.g.,
 as the beta@fitted_model from package topicmodels.
Or the actual top words (entries = words) per topic (columnnames = topics), e.g.,
as the output \code{terms(fitted_model, ...)} from topicmodesl pacakge or fitted_model$get_top_words(...) from text2vec package.}

\item{n}{Number of top words per topic to be used in the calculation. (if terms() or get_top_words() input is used,
the number needs to be the same as in these calls) - a better routine to check and adjust numbers will be implemented}
}
\value{
A \code{data.table} showing the coherence scores for different measures for each topic.
        For comparison of various models results need to be averaged over topics.
}
\description{
Calculation of various coherence measures for topic models based on LDA algorithm
}
\examples{

}
