% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_lm.R
\name{p_lm}
\alias{p_lm}
\title{Get p value of linear model}
\usage{
p_lm(model)
}
\arguments{
\item{model}{A linear model fitted via \code{lm}.}
}
\value{
The p-value of the model, i.e., the overall p-value of the F-statistic (not the p-value of the individual coefficients).
}
\description{
Get p value of linear model
}
\examples{

x = rnorm(42)
y = x + rnorm(42)
m = lm(y ~ x)
p_lm(m)
# [1] 2.783375e-09
}
