% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripl_to_sparse.R
\name{tripl_to_sparse}
\alias{tripl_to_sparse}
\title{Simple triple matrix to sparse matrix}
\usage{
tripl_to_sparse(simple_triplet_matrix, giveCsparse = TRUE)
}
\arguments{
\item{simple_triplet_matrix}{A simple_triplet_matrix as used in package slam.}

\item{giveCsparse}{By default \code {TRUE}. Set to \code{FALSE} if a \code{dgTMatrix} shall be returned instead of a \code{dgCMatrix}.}
}
\value{
\code{sparseMatrix} from package Matrix.
}
\description{
Simple triple matrix to sparse matrix
}
\examples{


}
